replaced Postgres Source strict-encrypt tests into new class
